+++ src/webviews/apps/shared/components/formatted-date.ts
@@ -2,6 +2,7 @@ import { html, LitElement } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import { formatDate, fromNow } from '../../../../system/date';
 import { dateConverter } from './converters/date-converter';
+import './overlays/tooltip';
 
 @customElement('formatted-date')
 export class FormattedDate extends LitElement {
@@ -16,8 +17,10 @@ export class FormattedDate extends LitElement {
 
 	override render() {
 		const formattedDate = formatDate(this.date, this.format ?? 'MMMM Do, YYYY h:mma');
-		return html`<time datetime="${this.date.toISOString()}" title="${formattedDate}"
-			>${this.dateStyle === 'relative' ? fromNow(this.date) : formattedDate}</time
+		return html`<gl-tooltip content="${formattedDate}"
+			><time datetime="${this.date.toISOString()}"
+				>${this.dateStyle === 'relative' ? fromNow(this.date) : formattedDate}</time
+			></gl-tooltip
 		>`;
 	}
 }